Refactor, main menu
