+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,7 @@
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -31,7 +32,7 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name=".Activities.MapInPagerActivity"
+            android:name=".activities.main.MapInPagerActivity"
             android:label="@string/PGGO_app_label">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -39,8 +40,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".Activities.BuildingDetailsActivity" />
-        <activity android:name=".Activities.FacultyDetailsActivity"></activity>
+        <activity android:name=".activities.app.BuildingDetailsActivity" />
+        <activity android:name=".activities.app.FacultyDetailsActivity" />
+        <activity android:name=".activities.fragments.OptionsActivity" />
     </application>
 
 </manifest>
