+++ app/src/main/assets/ETIfaculty.json
@@ -1,10 +1,18 @@
 {
   "name": "Elektroniki, telekomunikacji i informatyki",
-  "departments":[
-    {"name": "kaims" },
-    {"name": "kask" },
-    {"name": "kio" },
-    {"name": "kisi" },
+  "departments": [
+    {
+      "name": "kaims"
+    },
+    {
+      "name": "kask"
+    },
+    {
+      "name": "kio"
+    },
+    {
+      "name": "kisi"
+    },
     {
       "name": "ksg"
     }
