+++ app/src/main/assets/MECHANICZNY.json
@@ -1,6 +1,6 @@
 {
   "name": "Budynek wydziału mechanicznego",
-  "faculty" : "Mechaniczny",
+  "faculty": "Mechaniczny",
   "facultyTag": "MECHANICZNY",
   "picture": "mechaniczny_building.jpg",
   "description": "Wydział Mechaniczny – jeden z wydziałów Politechniki Gdańskiej. Wydział należy do najlepszych jednostek akademickich w Polsce, w 2017 uzyskał akredytację w kategorii A, nadaną przez Komitet Ewaluacji Jednostek Naukowych"
