+++ app/src/main/assets/NETI.json
@@ -1,7 +1,7 @@
 {
-	"name": "Nowe ETI",
-	"faculty" : "Elektroniki, telekomunikacji i informatyki",
-	"facultyTag": "ETI",
-	"picture": "eti_building.PNG",
-	"description": "Nowy budynek wydziału elektroniki, telekomunikacji i informatyki Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse pretium libero urna, scelerisque tincidunt nisl rutrum ultricies. Curabitur ornare purus justo, vitae semper risus varius non. Fusce aliquam eu enim eget mollis. Vivamus odio risus, ornare et luctus vitae, pretium sit amet arcu. In rhoncus purus in risus iaculis, ut mattis sem tincidunt. Phasellus id congue diam. Donec at augue augue. Suspendisse semper ex arcu, nec tempor erat scelerisque a. Nulla eu risus ac sem condimentum convallis. Sed viverra lorem ut orci egestas auctor. Curabitur vel dolor pellentesque, molestie quam nec, consequat sapien. Aenean ut ante et erat tempus ullamcorper at non diam. Maecenas scelerisque finibus massa et luctus. Nullam non volutpat purus. In viverra lorem sed semper hendrerit. Proin cursus velit vitae libero egestas tempor."
+  "name": "Nowe ETI",
+  "faculty": "Elektroniki, telekomunikacji i informatyki",
+  "facultyTag": "ETI",
+  "picture": "eti_building.PNG",
+  "description": "Nowy budynek wydziału elektroniki, telekomunikacji i informatyki Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse pretium libero urna, scelerisque tincidunt nisl rutrum ultricies. Curabitur ornare purus justo, vitae semper risus varius non. Fusce aliquam eu enim eget mollis. Vivamus odio risus, ornare et luctus vitae, pretium sit amet arcu. In rhoncus purus in risus iaculis, ut mattis sem tincidunt. Phasellus id congue diam. Donec at augue augue. Suspendisse semper ex arcu, nec tempor erat scelerisque a. Nulla eu risus ac sem condimentum convallis. Sed viverra lorem ut orci egestas auctor. Curabitur vel dolor pellentesque, molestie quam nec, consequat sapien. Aenean ut ante et erat tempus ullamcorper at non diam. Maecenas scelerisque finibus massa et luctus. Nullam non volutpat purus. In viverra lorem sed semper hendrerit. Proin cursus velit vitae libero egestas tempor."
 }
