--- app/src/main/java/pg/pgapp/Database/ModelType.java
@@ -1,5 +1,0 @@
-package pg.pgapp.Database;
-
-public enum ModelType {
-    BUILDING_DISPLAY, BUILDING, FACULTY, DEPARTMENT
-}
