+++ app/src/main/java/pg/pgapp/Initializer.java
@@ -17,70 +17,70 @@ import java.io.InputStream;
 import java.util.ArrayList;
 
 import lombok.NonNull;
-import pg.pgapp.Database.DatabaseConnector;
-import pg.pgapp.Models.BuildingDisplayModel;
+import pg.pgapp.database.DatabaseConnector;
+import pg.pgapp.models.BuildingDisplay;
 
 public class Initializer {
 
-    private final Context context;
+	private final Context context;
 
-    public Initializer(Context context) {
-        this.context = context;
-    }
+	public Initializer(Context context) {
+		this.context = context;
+	}
 
-    public void initialize(@NonNull final GoogleMap mMap) {
+	public void initialize(@NonNull final GoogleMap mMap) {
 
-        setBuildingsOnMap(mMap);
+		setBuildingsOnMap(mMap);
 
-        //pobranie mojej lokalizacji
-        if (ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED
-                && ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-            mMap.setMyLocationEnabled(true);
-        }
+		//pobranie mojej lokalizacji
+		if (ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED
+				&& ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+			mMap.setMyLocationEnabled(true);
+		}
 
-        LatLng noweEti = new LatLng(54.371648, 18.612357);
-        mMap.moveCamera(CameraUpdateFactory.newLatLng(noweEti)); //TODO: domyślnie będzie move to my location
-        mMap.animateCamera(CameraUpdateFactory.zoomTo(18));
-    }
+		LatLng noweEti = new LatLng(54.371648, 18.612357);
+		mMap.moveCamera(CameraUpdateFactory.newLatLng(noweEti)); //TODO: domyślnie będzie move to my location
+		mMap.animateCamera(CameraUpdateFactory.zoomTo(18));
+	}
 
-    private void setBuildingsOnMap(@NonNull final GoogleMap mMap) {
+	private void setBuildingsOnMap(@NonNull final GoogleMap mMap) {
         /*        //todo skonfigurowac ewentualne pobieranie danych z pliku lokalnie
-        Type listType = new TypeToken<ArrayList<BuildingDisplayModel>>() {
+        Type listType = new TypeToken<ArrayList<BuildingDisplay>>() {
         }.getType(); //potrzebne żeby wczytać listę obiektów z jsona
         Gson gson = new Gson();
-        ArrayList<BuildingDisplayModel> buildings = gson.fromJson(readDataFromFile("BuildingsConfiguration.json"), listType);
+        ArrayList<BuildingDisplay> buildings = gson.fromJson(readDataFromFile("BuildingsConfiguration.json"), listType);
         */
-        ArrayList<BuildingDisplayModel> buildings = new DatabaseConnector().getBuildingDisplays(0L);
-        buildings.forEach(
-                building -> {
-                    PolygonOptions buildingOptions = new PolygonOptions()
-                            .clickable(true)
-                            .strokeColor(Color.RED)
-                            .strokeWidth(2);
+		ArrayList<BuildingDisplay> buildings = new DatabaseConnector().getBuildingDisplays(0L);
+		buildings.forEach(
+				building -> {
+					PolygonOptions buildingOptions = new PolygonOptions()
+							.clickable(true)
+							.strokeColor(Color.RED)
+							.strokeWidth(2);
 
-                    building.getCoordinates().forEach(coordinate ->
-                            buildingOptions.add(new LatLng(coordinate.getLatitude(), coordinate.getLongitude())));
+					building.getCoordinates().forEach(coordinate ->
+							buildingOptions.add(new LatLng(coordinate.getLatitude(), coordinate.getLongitude())));
 
-                    Polygon polygon = mMap.addPolygon(buildingOptions);
-                    polygon.setClickable(true);
-                    polygon.setTag(building.getBuildingId());
-                }
-        );
-        mMap.setOnPolygonClickListener(new OnPolygonClickListener(context));
-    }
+					Polygon polygon = mMap.addPolygon(buildingOptions);
+					polygon.setClickable(true);
+					polygon.setTag(building.getBuildingId());
+				}
+		);
+		mMap.setOnPolygonClickListener(new OnPolygonClickListener(context));
+	}
 
-    private String readDataFromFile(String filename) {
-        String json = null;
-        try {
-            InputStream inputStream = context.getAssets().open(filename);
-            int size = inputStream.available();
-            byte[] buffer = new byte[size];
-            inputStream.read(buffer);
-            inputStream.close();
-            json = new String(buffer, "UTF-8");
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return json;
-    }
+	private String readDataFromFile(String filename) {
+		String json = null;
+		try {
+			InputStream inputStream = context.getAssets().open(filename);
+			int size = inputStream.available();
+			byte[] buffer = new byte[size];
+			inputStream.read(buffer);
+			inputStream.close();
+			json = new String(buffer, "UTF-8");
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		return json;
+	}
 }
