+++ app/src/main/java/pg/pgapp/OnPolygonClickListener.java
@@ -7,24 +7,20 @@ import android.util.Log;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.model.Polygon;
 
-import pg.pgapp.Activities.BuildingDetailsActivity;
-
-/**
- * Created by Ravo on 23.06.2018.
- */
+import pg.pgapp.activities.app.BuildingDetailsActivity;
 
 public class OnPolygonClickListener implements GoogleMap.OnPolygonClickListener {
-    private final Context _context;
+	private final Context _context;
 
-    public OnPolygonClickListener(Context context) {
-        _context = context;
-    }
+	public OnPolygonClickListener(Context context) {
+		_context = context;
+	}
 
-    @Override
-    public void onPolygonClick(Polygon polygon) {
-        Log.i("INFO", polygon.getTag().toString());
-        Intent intent = new Intent(_context, BuildingDetailsActivity.class);
-        intent.putExtra("TAG", polygon.getTag().toString());
-        _context.startActivity(intent);
-    }
+	@Override
+	public void onPolygonClick(Polygon polygon) {
+		Log.i("INFO", polygon.getTag().toString());
+		Intent intent = new Intent(_context, BuildingDetailsActivity.class);
+		intent.putExtra("TAG", polygon.getTag().toString());
+		_context.startActivity(intent);
+	}
 }
