+++ app/src/main/java/pg/pgapp/activities/fragments/ARFragment.java
@@ -1,0 +1,16 @@
+package pg.pgapp.activities.fragments;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import pg.pgapp.R;
+
+public class ARFragment extends Fragment {
+	@Override
+	public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle bundle) {
+		return inflater.inflate(R.layout.ar_placeholder, container, false);
+	}
+}
