+++ app/src/main/java/pg/pgapp/database/ModelType.java
@@ -1,0 +1,5 @@
+package pg.pgapp.database;
+
+public enum ModelType {
+	BUILDING_DISPLAY, BUILDING, FACULTY, DEPARTMENT
+}
