+++ app/src/main/java/pg/pgapp/models/BuildingDisplay.java
@@ -1,0 +1,24 @@
+package pg.pgapp.models;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class BuildingDisplay extends BaseModel {
+	private Long buildingId;
+	private List<Coordinate> coordinates;
+
+	@Getter
+	@Setter
+	@RequiredArgsConstructor
+	public class Coordinate {
+		private final Double latitude;
+		private final Double longitude;
+	}
+}
