+++ app/src/main/java/pg/pgapp/models/Faculty.java
@@ -1,0 +1,17 @@
+package pg.pgapp.models;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class Faculty extends BaseModel {
+	private List<String> departmentsNames;
+
+	public Faculty(long id, String name, String tag, List<String> departmentsNames) {
+		super(id, name, tag);
+		this.departmentsNames = departmentsNames;
+	}
+}
