+++ app/src/main/res/layout/activity_maps.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/map"
     android:name="com.google.android.gms.maps.SupportMapFragment"
+    class="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="pg.pgapp.ETIMapFragment"
-    class="com.google.android.gms.maps.SupportMapFragment"/>
+    tools:context="pg.pgapp.activities.fragments.ETIMapFragment" />
