+++ app/src/main/res/layout/main_menu.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<GridLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="horizontal"
+    android:background="@color/colorPgPrimary">
+
+    <TextView xmlns:android="http://schemas.android.com/apk/res/android"
+        android:layout_width="match_parent"
+        android:layout_height="83dp"
+        android:layout_column="0"
+        android:layout_row="1"
+        android:gravity="center"
+        android:textColor="@color/colorWhite"
+        android:text="@string/main_menu" />
+
+    <Button
+        android:id="@+id/runOptionsButton"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_column="0"
+        android:layout_row="3"
+        android:background="@color/colorPgSecondary"
+        android:gravity="center"
+        android:textColor="@color/colorWhite"
+        android:text="@string/options_button" />
+
+    <ImageView
+        android:id="@+id/imageView2"
+        android:layout_width="398dp"
+        android:layout_height="233dp"
+        android:layout_column="0"
+        android:layout_row="0"
+        app:srcCompat="@mipmap/logo" />
+</GridLayout>
