+++ app/src/main/res/values-night/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="options_button">Ustawienia</string>
+</resources>
