+++ app/src/main/res/values/theme.xml
@@ -1,11 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+
     <style name="FontSizeSmall">
         <item name="android:textSize">12sp</item>
     </style>
+
     <style name="FontSizeMedium">
         <item name="android:textSize">16sp</item>
     </style>
+
     <style name="FontSizeLarge">
         <item name="android:textSize">20sp</item>
     </style>
